Add the option to use generated or actual data.
